move ETH check to command + fix UI limit check
